ft(location) - Add location to each move.
- Add .vscode to gitignore.
- Add position array to history obj in state.
- Display location in format row & col.
- Factor out utils into folder.
