+++ src/utils/calculateWinner.js
