+++ src/utils/getLocation.js
@@ -1,0 +1,43 @@
+function getLocation(i) {
+    let locationObj = {
+      0: {
+        row: 0,
+        col: 0,
+      },
+      1: {
+        row: 0,
+        col: 1,
+      },
+      2: {
+        row: 0,
+        col: 2,
+      },
+      3: {
+        row: 1,
+        col: 0,
+      },
+      4: {
+        row: 1,
+        col: 1,
+      },
+      5: {
+        row: 1,
+        col: 2,
+      },
+      6: {
+        row: 2,
+        col: 0,
+      },
+      7: {
+        row: 2,
+        col: 1,
+      },
+      8: {
+        row: 2,
+        col: 2,
+      }
+    }
+    return locationObj[i];
+  }
+
+  export default getLocation;
